+++ tests/sql_test_files/it_works_basic_auth_password.sql
@@ -1,0 +1,5 @@
+select 'text' as component,
+    case sqlpage.basic_auth_password()
+        when 'test' then 'It works !'
+        else 'error: ' || coalesce(sqlpage.basic_auth_password(), 'NULL')
+    end as contents;